+++ src/test/kotlin/lotto/model/LottoMachineTest.kt
@@ -1,6 +1,5 @@
-package lotto
+package lotto.model
 
-import lotto.model.LottoMachine
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
