+++ src/test/kotlin/lotto/model/LottoNumberTest.kt
@@ -1,6 +1,5 @@
-package lotto
+package lotto.model
 
-import lotto.model.LottoNumber
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
