+++ src/test/kotlin/lotto/model/LottoResultTest.kt
@@ -1,7 +1,5 @@
-package lotto
+package lotto.model
 
-import lotto.model.LottoResult
-import lotto.model.Rank
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
