+++ src/test/kotlin/lotto/model/LottoScannerTest.kt
@@ -1,9 +1,5 @@
-package lotto
+package lotto.model
 
-import lotto.model.Lotto
-import lotto.model.LottoNumber
-import lotto.model.LottoScanner
-import lotto.model.Rank
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
